Color Game in Python
In this game, the player has to enter the color of the word that appears on the screen, and hence the score increases by one, the total time to play this game is 1 minute. Interface will display name of different colors in different colors. Player has to identify the color and enter the correct color name to win the game.