Changed the input portion of code.
I went through the program and noticed that it was having predefined inputs, that is, the elements of the array to be sorted. I changed the input portion of the code and made it in such a way that the user can define the size of the array he needs and the elements of the array too as per his/her requirement.